Use class instance varaible @cached_color
if we use class variable, @@cached_colors is available in three color
palettes. In this case, there will be a nasty bug. For example, green
in Rainbow palette is [0, 128, 0], if we accessed to it and then want
to have a look at the green color in Primary palette, we will still
get the [0, 128, 0], not the real [0, 255, 0] because of the cache.
Use @cached_color instead of @@cached_color can avoid this problem.